requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-RESTful==0.3.9
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymssql==2.2.7
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
sgf==0.5
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
Werkzeug==2.2.3
